Refactor code to fix dependency cycle
