+++ src/gameAPI.js
@@ -1,3 +1,5 @@
+import myScoreBoard from './scoreBoard';
+
 const fetch = require('node-fetch');
 
 const requestURL = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/';
@@ -14,6 +16,43 @@ const scoresBoard = () => {
     },
   })
     .then((response) => response.json());
+};
+
+const gameURL = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/0SnEe00QRycoJQ0WLR02/scores/';
+
+export const getGameScores = async () => {
+  try {
+    await fetch(gameURL, {
+      mode: 'cors',
+      method: 'GET'
+    })
+    .then((response) => response.json())
+    .then((data) => myScoreBoard.assignScores(data.result));
+  } catch(error) {
+    myScoreBoard.assignScores(null);
+  }
+};
+
+export const addGameScore = async (name, points) => {
+try {
+  processingMessage.classList.remove('hidden')
+  refreshButton.disabled = true;
+  submitButton.disabled = true;
+  await fetch(gameURL, {
+    mode: 'cors',
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({"user": name, "score": points }),
+  }).then((response) => {
+    console.log(response.json());
+  }).then(() => {
+    processingMessage.classList.add('hidden');
+    refreshButton.disabled = false;
+    submitButton.disabled = false;
+  });
+} catch(error) {
+  console.log(error);
+}
 };
 
 export default scoresBoard;
